+++ google-cloud/container-linux/kubernetes/controllers.tf
@@ -1,6 +1,6 @@
 # Discrete DNS records for each controller's private IPv4 for etcd usage
 resource "google_dns_record_set" "etcds" {
-  count = "${var.count}"
+  count = "${var.controller_count}"
 
   # DNS Zone name where record should be created
   managed_zone = "${var.dns_zone_name}"
@@ -21,11 +21,11 @@ data "google_compute_zones" "all" {
 
 # Controller instances
 resource "google_compute_instance" "controllers" {
-  count = "${var.count}"
+  count = "${var.controller_count}"
 
   name         = "${var.cluster_name}-controller-${count.index}"
   zone         = "${element(data.google_compute_zones.all.names, count.index)}"
-  machine_type = "${var.machine_type}"
+  machine_type = "${var.controller_type}"
 
   metadata {
     user-data = "${element(data.ct_config.controller_ign.*.rendered, count.index)}"
@@ -41,7 +41,7 @@ resource "google_compute_instance" "controllers" {
   }
 
   network_interface {
-    network = "${var.network}"
+    network = "${google_compute_network.network.name}"
 
     # Ephemeral external IP
     access_config = {}
@@ -51,9 +51,13 @@ resource "google_compute_instance" "controllers" {
   tags           = ["${var.cluster_name}-controller"]
 }
 
+locals {
+  controllers_ipv4_public = ["${google_compute_instance.controllers.*.network_interface.0.access_config.0.assigned_nat_ip}"]
+}
+
 # Controller Container Linux Config
 data "template_file" "controller_config" {
-  count = "${var.count}"
+  count = "${var.controller_count}"
 
   template = "${file("${path.module}/cl/controller.yaml.tmpl")}"
 
@@ -65,7 +69,7 @@ data "template_file" "controller_config" {
     # etcd0=https://cluster-etcd0.example.com,etcd1=https://cluster-etcd1.example.com,...
     etcd_initial_cluster = "${join(",", formatlist("%s=https://%s:2380", null_resource.repeat.*.triggers.name, null_resource.repeat.*.triggers.domain))}"
 
-    kubeconfig            = "${indent(10, var.kubeconfig)}"
+    kubeconfig            = "${indent(10, module.bootkube.kubeconfig)}"
     ssh_authorized_key    = "${var.ssh_authorized_key}"
     k8s_dns_service_ip    = "${cidrhost(var.service_cidr, 10)}"
     cluster_domain_suffix = "${var.cluster_domain_suffix}"
@@ -75,7 +79,7 @@ data "template_file" "controller_config" {
 # Horrible hack to generate a Terraform list of a desired length without dependencies.
 # Ideal ${repeat("etcd", 3) -> ["etcd", "etcd", "etcd"]}
 resource null_resource "repeat" {
-  count = "${var.count}"
+  count = "${var.controller_count}"
 
   triggers {
     name   = "etcd${count.index}"
@@ -84,8 +88,8 @@ resource null_resource "repeat" {
 }
 
 data "ct_config" "controller_ign" {
-  count        = "${var.count}"
+  count        = "${var.controller_count}"
   content      = "${element(data.template_file.controller_config.*.rendered, count.index)}"
   pretty_print = false
-  snippets     = ["${var.clc_snippets}"]
+  snippets     = ["${var.controller_clc_snippets}"]
 }